기존 체형 가져오기, 새로운 체형 입력 화면 추가
