+++ .idea/misc.xml
@@ -18,7 +18,9 @@
         <entry key="..\:/Users/msi/Desktop/Class_Project/app/src/main/res/drawable/squre_border_black.xml" value="0.3075" />
         <entry key="..\:/Users/msi/Desktop/Class_Project/app/src/main/res/drawable/straight_neck.xml" value="0.3075" />
         <entry key="..\:/Users/msi/Desktop/Class_Project/app/src/main/res/layout/activity_main.xml" value="0.33" />
+        <entry key="..\:/Users/msi/Desktop/Class_Project/app/src/main/res/layout/activity_select.xml" value="0.36614583333333334" />
         <entry key="..\:/Users/msi/Desktop/Class_Project/app/src/main/res/layout/result_activity.xml" value="0.36614583333333334" />
+        <entry key="..\:/Users/msi/Desktop/Class_Project/app/src/main/res/layout/test.xml" value="0.36614583333333334" />
         <entry key="..\:/Users/msi/Desktop/Class_Project/app/src/main/res/mipmap-anydpi-v26/ic_launcher.xml" value="0.3075" />
       </map>
     </option>
